+++ examples/provider/dwd/observation/dwd_obs_values_sql.py
@@ -28,7 +28,7 @@ log = logging.getLogger()
 
 def values_sql_example():
     """Retrieve temperature data by DWD and filter by sql statement."""
-    settings = Settings(ts_shape="long", ts_humanize=True, ts_si_units=False)
+    settings = Settings(ts_shape="long", ts_humanize=True, ts_convert_units=False)
 
     request = DwdObservationRequest(
         parameters=("hourly", "temperature_air"),