+++ wetterdienst/core/timeseries/request.py
@@ -164,7 +164,7 @@ class TimeseriesRequest(Core):
         self.humanize = settings.ts_humanize
         self.shape = settings.ts_shape
         self.tidy = 1 if self.shape == "long" else 0
-        self.si_units = settings.ts_si_units
+        self.convert_units = settings.ts_convert_units
 
         self.drop_nulls = self.tidy and settings.ts_drop_nulls
         self.complete = not self.drop_nulls and settings.ts_complete
@@ -315,8 +315,8 @@ class TimeseriesRequest(Core):
                         {
                             "name": parameter.name,
                             "name_original": parameter.name_original,
-                            "unit": parameter.unit,
-                            "unit_original": parameter.unit_original,
+                            "unit": parameter.unit_type,
+                            "unit_original": parameter.unit,
                         }
                     )
             if not data[resolution.name]: