+++ wetterdienst/settings.py
@@ -11,6 +11,7 @@ import platformdirs
 from pydantic import Field, field_validator, model_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
+from wetterdienst.core.timeseries.unit import UnitConverter
 from wetterdienst.metadata.parameter import Parameter
 
 log = logging.getLogger(__name__)
@@ -26,7 +27,8 @@ class Settings(BaseSettings):
     fsspec_client_kwargs: dict = Field(default_factory=dict)
     ts_humanize: bool = True
     ts_shape: Literal["wide", "long"] = "long"
-    ts_si_units: bool = True
+    ts_convert_units: bool = True
+    ts_unit_targets: dict[str, str] = Field(default_factory=dict)
     ts_skip_empty: bool = False
     ts_skip_threshold: float = 0.95
     ts_skip_criteria: Literal["min", "mean", "max"] = "min"
@@ -39,6 +41,13 @@ class Settings(BaseSettings):
         }
     )
     ts_interpolation_use_nearby_station_distance: float = 1.0
+
+    @field_validator("ts_unit_targets", mode="after")
+    @classmethod
+    def validate_ts_unit_targets(cls, values):
+        if not values.keys() <= UnitConverter.targets.keys():
+            raise ValueError(f"Invalid unit targets: one of {set(values.keys())} not in {set(UnitConverter.targets.keys())}")
+        return values
 
     # make ts_interpolation_station_distance update but not replace the default values
     @field_validator("ts_interpolation_station_distance", mode="before")