+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Feature
 - Add logo for restapi
+- Add dedicated unit converter
 
 ## 0.99.0 - 2024-12-30